txn: move actions/shared fns
