+++ src/storage/txn/actions/mod.rs
@@ -5,8 +5,6 @@
 //! eg. [`MvccTxn::load_lock`], [`MvccTxn::put_write`], which are methods on [`MvccTxn`],
 //! for archiving a certain target
 
-pub(crate) mod shared;
-
 pub mod acquire_pessimistic_lock;
 pub mod check_data_constraint;
 pub mod check_txn_status;
